requirements.txt
boto3>=1.34
requests>=2.31
pyarrow>=15.0
pyiceberg[s3,rest]>=0.6
pyyaml>=6.0
python-dotenv>=1.0
